requirements.txt
# Core dependencies for web scraping
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Progress bars and UI
tqdm>=4.64.0

# Google Gemini API
google-generativeai>=0.3.0

# Data models and validation
pydantic>=2.0.0

# Configuration management
pyyaml>=6.0.0

# Environment variables
python-dotenv>=1.0.0

#openai
openai>=1.0.0

# statistics
pandas>=2.0.0

# ruff
ruff>=0.10.0